requirements.txt
Django==5.0.7
django-ninja==1.2.2
pytz==2024.1